Bugfix : Null Json Obfuscation
Fix the issue where we were unable to run the app in release, the issue was due to the obfuscation made when deploying an app on release which caused the json inside the p2p library to be null.
Added a R8/Proguard config in the android release to remove the obfuscation as serialising the whole p2p library would take too long.